added search results page (/search/:id), fetch is done there (in route) now instead of in /utils/fetchProduct. public/procuctSearch now simply redirects the user when they search something
